requirements.txt
transformers
sentencepiece
datasets
pathlib
openpyxl
matplotlib
seaborn
xgboost
accelerate
scikit-learn
numpy
lightning
pytorch-lightning
sentence-transformers